refactor(core): remove response data construction helper from libs
